(feat) Allow toolbar icons to be disabled
A toolbar icon in the disabled mode will not accept hotkeys (e.g. ctrl+b
for bold) and with the provided css will be greyed out.

The css class wysiwyg-toolbar-icon-disabled is added to the element when
disabled and can be used to e.g. hide the element when disabled.

To disable an icon, simply call
$element.trigger('wysiwyg-toolbar-icon-disable'), and
$element.trigger('wysiwyg-toolbar-icon-enable') to enable it again.